improved potential duplicates flagging

+++ utils/handle_duplicates.py
@@ -1,6 +1,7 @@
 from .filename_reader import remove_extension, extract_extension
 import json
 import os
+import filecmp
 
 import nibabel as nib
 import numpy as np
@@ -12,9 +13,58 @@ from dropbox.files import WriteMode
 from dropbox.exceptions import ApiError, AuthError
 from tqdm import tqdm
 
+def flag_same_new_paths(file_infos_path, flagged_path):
+    """
+    Saves a json in `flagged_path` flagging different files that may have the same new path
+
+    **Overwrites any existing file in `flagged_path`**
+
+
+    Parameters
+    --------
+        file_infos_path : str,
+            path to fileinfos
+        flagged_path : str,
+            the path where the flags will be saved (usually of the type `/file_infos/subdir/same_new_paths-[n].json`)
+    
+    Saves
+    --------
+        flagged_path, json file
+            flagged files can be found as follows: flagged_dict[file1] = list(possible duplicate files, file1 included)
+
+    """
+    with open(file_infos_path, 'r') as f:
+        file_infos = json.load(f)
+    
+    visited = []
+
+    out_dict = {}
+
+    for file1 in file_infos.keys():
+        file1_duplicates_list = []
+        
+
+        if file1 not in visited:            
+            for file2 in file_infos.keys():
+
+                if file2 not in visited:
+                    condition1 = file_infos[file1]['new_path']==file_infos[file2]['new_path']
+                    if condition1:
+                        file1_duplicates_list.append(file2)
+                        visited.append(file2)
+            visited.append(file1)
+        if len(file1_duplicates_list)>1:
+            out_dict[file1] = file1_duplicates_list
+        del file1_duplicates_list
+
+    out_dirs = '/'.join(flagged_path.split('/')[:-1])
+    os.makedirs(out_dirs, exist_ok=True)
+    with open(flagged_path, 'w') as f:
+        json.dump(out_dict, f, indent=4)
+
 def flag_potential_duplicates(file_infos_path, flagged_path):
     """
-    Saves a json in `flagged_path` flagging different files that may have the same name after sorting
+    Saves a json in `flagged_path` flagging different files that may be duplicates
 
     **Overwrites any existing file in `flagged_path`**
 
@@ -41,12 +91,48 @@ def flag_potential_duplicates(file_infos_path, flagged_path):
 
     for file1 in file_infos.keys():
         file1_duplicates_list = []
-        if file1 not in visited:
-            
+        filename1 = file_infos[file1]['old_path'].split('/')[-1]
+        type1 = file_infos[file1]['type']
+        id1 = file_infos[file1]['id']
+        extension1 = extract_extension(filename1)
+        try:
+            seg_info1 = file_infos[file1]['seg_info']
+        except:
+            seg_info1 = ''
+        try:
+            func_task1 = file_infos[file1]['func_task']
+        except:
+            func_task1 = ''
+        try:
+            func_info1 = file_infos[file1]['func_info']
+        except:
+            func_info1 = ''
 
+        if file1 not in visited:            
             for file2 in file_infos.keys():
+                filename2 = file_infos[file2]['old_path'].split('/')[-1]
+                type2 = file_infos[file2]['type']
+                id2 = file_infos[file2]['id']
+                extension2 = extract_extension(filename2)
+                try:
+                    seg_info2 = file_infos[file2]['seg_info']
+                except:
+                    seg_info2 = ''
+                try:
+                    func_task2 = file_infos[file2]['func_task']
+                except:
+                    func_task2 = ''
+                try:
+                    func_info2 = file_infos[file2]['func_info']
+                except:
+                    func_info2 = ''
+
                 if file2 not in visited:
-                    if file_infos[file1]['new_path']==file_infos[file2]['new_path']:
+                    condition1 = file_infos[file1]['new_path']==file_infos[file2]['new_path']
+                    condition2 = filename1 == filename2
+                    condition3 = (type1==type2) and (id1==id2) and (seg_info1==seg_info2) and (func_task1==func_task2) and (func_info1==func_info2)
+
+                    if (condition1 or condition2 or condition3) and (extension1==extension2) and (type1 not in ['code', 'misc', 'modelling']):
                         file1_duplicates_list.append(file2)
                         visited.append(file2)
             visited.append(file1)
@@ -137,39 +223,63 @@ def compare_potential_duplicates(flagged_path, actual_duplicates_path, TOKEN, ve
         for file1_idx in range(n-1):
             from_path1 = '/source' + flagged_duplicates[key][file1_idx]
             from_path1_without_source = flagged_duplicates[key][file1_idx]
-            to_path1 = 'tmp_dir/file1.nii.gz'
+            to_path1 = 'tmp_dir/file1' + extract_extension(from_path1)
 
-            actual_duplicates[from_path1_without_source] = [from_path1_without_source]
-            
-            dbx.files_download_to_file(to_path1, from_path1)
+            try:
+                dbx.files_download_to_file(to_path1, from_path1)
+                downloaded1 = True
+            except:
+                downloaded1 = False
+                print('failed to download ' + from_path1)
 
-            for file2_idx in range(file1_idx+1, n):
-                from_path2 = '/source' + flagged_duplicates[key][file2_idx]
-                from_path2_without_source = flagged_duplicates[key][file2_idx]
-                to_path2 = 'tmp_dir/file2.nii.gz'
 
-                if debug:    
-                    print(from_path1)
-                    print(from_path2)
+            if downloaded1:
+                actual_duplicates[from_path1_without_source] = [from_path1_without_source]
 
-                
-                dbx.files_download_to_file(to_path2, from_path2)
+                for file2_idx in range(file1_idx+1, n):
+                    from_path2 = '/source' + flagged_duplicates[key][file2_idx]
+                    from_path2_without_source = flagged_duplicates[key][file2_idx]
+                    to_path2 = 'tmp_dir/file2' + extract_extension(from_path2)
 
-                if extract_extension(from_path1) == '.nii.gz' and extract_extension(from_path2) == '.nii.gz':
+                    if debug:    
+                        print(from_path1)
+                        print(from_path2)
+                    
+                    try:
+                        dbx.files_download_to_file(to_path2, from_path2)
+                        downloaded2 = True
+                    except:
+                        downloaded2 = False
+                        print('failed to download ' + from_path2)
+                    
+                    if downloaded2:
 
-                    img1 = nib.load(to_path1)
-                    img2 = nib.load(to_path2)
+                        if extract_extension(from_path1) == '.nii.gz' and extract_extension(from_path2) == '.nii.gz':
 
-                    data1 = img1.get_fdata()
-                    data2 = img2.get_fdata()
+                            img1 = nib.load(to_path1)
+                            img2 = nib.load(to_path2)
 
-                    if data1.shape == data2.shape:
-                        if (data1==data2).all():
-                            if debug:
+                            data1 = img1.get_fdata()
+                            data2 = img2.get_fdata()
+
+                            if data1.shape == data2.shape:
+                                if (data1==data2).all():
+                                    if debug:
+                                        print(from_path1 + '\n    is the same as: \n' + from_path2)
+                                    actual_duplicates[from_path1_without_source].append(from_path2_without_source)
+                        else:
+                            try:
+                                comp = filecmp.cmp(to_path1, to_path2, shallow=False)
+                            except:
+                                comp=False
+                                print(from_path1 + ' not compared')
+
+                            if comp:
                                 print(from_path1 + '\n    is the same as: \n' + from_path2)
-                            actual_duplicates[from_path1_without_source].append(from_path2_without_source)
-                os.remove(to_path2)
-            os.remove(to_path1)
+                                actual_duplicates[from_path1[len('/source'):]].append(from_path2[len('/source'):])
+
+                        os.remove(to_path2)
+                os.remove(to_path1)
 
     with open(actual_duplicates_path, 'w') as f:
         json.dump(actual_duplicates, f, indent=4)
